requirements.txt
pandas~=2.2.2
CouchDB~=1.2
couchbase~=4.3.0
Faker~=13.16.0